+++ forms/src/misc/formModel/decorators/formArrayChild.decorator.ts
@@ -1,0 +1,15 @@
+import {FormArray, FormControl, FormGroup} from '@angular/forms';
+
+import {ModelPropertyMetadata} from './modelPropertyMetadata.decorator';
+
+/**
+ * Defines child type for FormArray
+ * @param childType - Type of child that should be created inside FormArray
+ */
+export function FormArrayChild(childType: typeof FormGroup|typeof FormArray|typeof FormControl): PropertyDecorator
+{
+    return ModelPropertyMetadata(
+    {
+        childType: childType
+    });
+}