+++ forms/src/misc/formModel/decorators/index.ts
@@ -2,6 +2,7 @@ export * from './asyncValidator.decorator';
 export * from './controlOptions.decorator';
 export * from './disabled.decorator';
 export * from './email.decorator';
+export * from './formArrayChild.decorator';
 export * from './formArrayProperty.decorator';
 export * from './formGroupAsyncValidator.decorator';
 export * from './formGroupProperty.decorator';